fix(app2): barchart component
