+++ app2/src/lib/components/model/BarChart.svelte
@@ -1,0 +1,109 @@
+<script lang="ts">
+  import { Option } from "effect"
+  import DateTimeComponent from "$lib/components/ui/DateTimeComponent.svelte"
+  import type { DailyTransfer } from "$lib/schema/statistics"
+  import ErrorComponent from "$lib/components/model/ErrorComponent.svelte"
+  import type { FetchDecodeGraphqlError } from "$lib/utils/queries"
+
+  type Props = {
+    data: Option.Option<Array<DailyTransfer>>
+    error: Option.Option<FetchDecodeGraphqlError>
+    class?: string
+  }
+
+  const { data, error, class: className = "" } = $props()
+
+  // Format large numbers with commas (used for chart tooltips)
+  function formatNumber(num: string | number): string {
+    return Number(num).toLocaleString()
+  }
+
+  // Derived values for chart data
+  const reversedDailyTransfers = $derived(
+    Option.isSome(data) ? [...data.value].reverse() : []
+  )
+
+  const maxCount = $derived(
+    Option.isSome(data) ? Math.max(...data.value.map(d => d.count)) : 0
+  )
+
+  // Calculate nice round numbers for y-axis labels
+  const yLabels = $derived(() => {
+    if (maxCount <= 0) return [0, 0, 0, 0, 0]
+
+    // Find a nice round maximum that's at least as large as maxCount
+    const magnitude = 10 ** Math.floor(Math.log10(maxCount))
+    const roundedMax = Math.ceil(maxCount / magnitude) * magnitude
+
+    // Create evenly spaced labels
+    return [
+      0,
+      Math.round(roundedMax / 4),
+      Math.round(roundedMax / 2),
+      Math.round((roundedMax * 3) / 4),
+      roundedMax
+    ]
+  })
+
+  // Calculate bar heights as percentages
+  const barHeights = $derived(
+    reversedDailyTransfers.map(day => ({
+      ...day,
+      heightPercent: maxCount > 0 ? Math.max((day.count / maxCount) * 100, 1) : 1
+    }))
+  )
+
+  // Get labels for x-axis (first, middle, last)
+  const xAxisLabels = $derived(
+    reversedDailyTransfers.filter(
+      (_, i, arr) => i === 0 || i === Math.floor(arr.length / 2) || i === arr.length - 1
+    )
+  )
+</script>
+
+{#if Option.isSome(data) && maxCount > 0}
+  <!-- Chart container -->
+  <div class="h-80 relative chart-container {className}">
+    <!-- Grid lines -->
+    <div class="absolute left-0 right-0 top-0 bottom-0 flex flex-col justify-between">
+      {#each Array(5) as _, i}
+        <div class="border-t first:border-0 border-zinc-200 dark:border-zinc-900 w-full h-0"></div>
+      {/each}
+    </div>
+    
+    <!-- Bars -->
+    <div class="absolute left-0 right-0 top-0 bottom-0 pt-1 px-4">
+      <div class="flex h-full gap-[1px] sm:gap-[2px] md:gap-1 items-end" style="min-height: 12rem;">
+        {#each barHeights as day, i}
+          <div class="flex flex-col flex-1 group size-full justify-end hover:opacity-100">
+            <div class="w-full size-full flex items-end">
+              <div 
+                class="relative w-full bg-white rounded-t transition-all duration-300 bar"
+                style="height: {day.heightPercent}%; min-height: 1px;"
+              >
+                <div class="absolute pointer-events-none bottom-full mb-2 left-1/2 transform -translate-x-1/2 bg-zinc-950 border-zinc-900 border text-white dark:text-white px-2 py-1 rounded text-xs opacity-0 group-hover:opacity-100 transition-opacity whitespace-nowrap z-10">
+                  <div>{formatNumber(day.count)}</div> <DateTimeComponent value={day.day} showTime={false} />
+                </div>
+              </div>
+            </div>
+          </div>
+        {/each}
+      </div>
+    </div>
+  </div>
+{:else if Option.isSome(error)}
+  <ErrorComponent error={error.value} />
+{:else}
+  <div class="h-64 animate-pulse bg-zinc-200 dark:bg-zinc-700 rounded"></div>
+{/if}
+
+<style>
+  /* Style for chart bars - make non-hovered bars darker when any bar is hovered */
+  :global(.chart-container:hover .flex-1) {
+    opacity: 0.3;
+  }
+  
+  :global(.chart-container .flex-1:hover) {
+    opacity: 1 !important;
+  }
+</style>
